Make Element Desktop send telegram sticker info (#4)
Element Desktop removes additional properties in the sticker info, but
not the media info. Element Android removes them in both currently. This
commit moves the net.maunium.telegram.sticker property to the media
info. This can be combined with a change to mautrix-telegram that
bridges telegram-imported matrix stickers to matrix stickers.

Co-authored-by: Charlotte 🦝 Delenk <lotte@chir.rs>